genesis_replicator/backup_recovery/backup_manager.py
"""
Backup Manager for handling system-wide backups.
"""
import asyncio
import json
import logging
import os
from datetime import datetime
from pathlib import Path
from typing import Dict, List, Optional

logger = logging.getLogger(__name__)

class BackupManager:
    """Manages system-wide backups including state, configuration, and models."""

    def __init__(self, backup_dir: str = "backups"):
        """Initialize the backup manager.

        Args:
            backup_dir: Directory to store backups
        """
        self.backup_dir = Path(backup_dir)
        self.backup_dir.mkdir(parents=True, exist_ok=True)
        self._backup_lock = asyncio.Lock()

    async def create_backup(self, components: List[str]) -> str:
        """Create a new backup of specified components.

        Args:
            components: List of components to backup

        Returns:
            Backup ID
        """
        async with self._backup_lock:
            backup_id = datetime.now().strftime("%Y%m%d_%H%M%S")
            backup_path = self.backup_dir / backup_id
            backup_path.mkdir()

            try:
                for component in components:
                    await self._backup_component(component, backup_path)

                # Create backup metadata
                metadata = {
                    "backup_id": backup_id,
                    "timestamp": datetime.now().isoformat(),
                    "components": components,
                    "status": "completed"
                }

                with open(backup_path / "metadata.json", "w") as f:
                    json.dump(metadata, f, indent=2)

                return backup_id

            except Exception as e:
                logger.error(f"Backup failed: {str(e)}")
                raise

    async def restore_backup(self, backup_id: str, components: Optional[List[str]] = None) -> bool:
        """Restore system from a backup.

        Args:
            backup_id: ID of the backup to restore
            components: Optional list of specific components to restore

        Returns:
            Success status
        """
        backup_path = self.backup_dir / backup_id
        if not backup_path.exists():
            raise ValueError(f"Backup {backup_id} not found")

        try:
            # Read backup metadata
            with open(backup_path / "metadata.json") as f:
                metadata = json.load(f)

            # Determine components to restore
            restore_components = components or metadata["components"]

            # Verify all requested components exist in backup
            for component in restore_components:
                if component not in metadata["components"]:
                    raise ValueError(f"Component {component} not found in backup")

            # Perform restoration
            for component in restore_components:
                await self._restore_component(component, backup_path)

            return True

        except Exception as e:
            logger.error(f"Restore failed: {str(e)}")
            raise

    async def list_backups(self) -> List[Dict]:
        """List all available backups.

        Returns:
            List of backup metadata
        """
        backups = []
        for backup_dir in self.backup_dir.iterdir():
            if backup_dir.is_dir():
                try:
                    with open(backup_dir / "metadata.json") as f:
                        metadata = json.load(f)
                        backups.append(metadata)
                except Exception as e:
                    logger.warning(f"Failed to read backup metadata: {str(e)}")

        return sorted(backups, key=lambda x: x["timestamp"], reverse=True)

    async def _backup_component(self, component: str, backup_path: Path) -> None:
        """Backup a specific component.

        Args:
            component: Component name
            backup_path: Path to store backup
        """
        component_path = backup_path / component
        component_path.mkdir()

        # Component-specific backup logic
        if component == "state":
            await self._backup_state(component_path)
        elif component == "config":
            await self._backup_config(component_path)
        elif component == "models":
            await self._backup_models(component_path)
        else:
            raise ValueError(f"Unknown component: {component}")

    async def _restore_component(self, component: str, backup_path: Path) -> None:
        """Restore a specific component.

        Args:
            component: Component name
            backup_path: Path to backup
        """
        component_path = backup_path / component
        if not component_path.exists():
            raise ValueError(f"Component {component} not found in backup")

        # Component-specific restore logic
        if component == "state":
            await self._restore_state(component_path)
        elif component == "config":
            await self._restore_config(component_path)
        elif component == "models":
            await self._restore_models(component_path)
        else:
            raise ValueError(f"Unknown component: {component}")

    async def _backup_state(self, path: Path) -> None:
        """Backup system state."""
        # Implementation for state backup
        pass

    async def _backup_config(self, path: Path) -> None:
        """Backup configuration."""
        # Implementation for config backup
        pass

    async def _backup_models(self, path: Path) -> None:
        """Backup AI models."""
        # Implementation for model backup
        pass

    async def _restore_state(self, path: Path) -> None:
        """Restore system state."""
        # Implementation for state restore
        pass

    async def _restore_config(self, path: Path) -> None:
        """Restore configuration."""
        # Implementation for config restore
        pass

    async def _restore_models(self, path: Path) -> None:
        """Restore AI models."""
        # Implementation for model restore
        pass
